% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_public.R
\docType{methods}
\name{accInfo}
\alias{accInfo}
\alias{accInfo,Commons-method}
\title{Display meta-info of dbGaP objects}
\usage{
accInfo(object, acc, ...)

\S4method{accInfo}{Commons}(object, acc, ..., display = TRUE)
}
\arguments{
\item{object}{Commons class object.}

\item{acc}{a character string. The dbGaP study, dataset, or variable accession.}

\item{...}{There are optional arguments.}

\item{display}{a logical value. (optional) If TRUE (default), console displays the meta-info. If FALSE, not display.}

\item{dataStudyOnly}{a logical value. When TRUE (default), only downloads the dataset and variable metadata of the stdudies that have data files in the project directory.  When FALSE, downloads the dataset and variable metadata of all dbGaP released studies, regardless the actual phenotype data files of the studies are downloaded or not.}
}
\value{
a data frame. (invisible) The meta-info of the input object.
}
\description{
The method displays the meta-info of the dbGaP objects given an object accession. The object can be a dbGaP study, dataset, or variable.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: returns dataset accession
}}

\examples{
\dontrun{

c <- Commons()
df <- accInfo(c, acc = 'phs000001.v3.p1')
# or
df <- accInfo(c, acc = 'pht000824.v5')
# or
df <- accInfo(c, acc = 'phv00000090.v2')     # numeric with units
# or
df <- accInfo(c, acc = 'phv00000122.v2')     # categorical with code-value 
}
}
