% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{buildVarCodeValCombo}
\alias{buildVarCodeValCombo}
\alias{buildVarCodeValCombo,Commons,character-method}
\title{(internal) Compose variable code-value combo string.}
\usage{
buildVarCodeValCombo(object, catVarAcc, ...)

\S4method{buildVarCodeValCombo}{Commons,character}(object, catVarAcc, ...,
  phsAcc = "", varCodeValDF = data.frame())
}
\arguments{
\item{object}{Commons class object.}

\item{catVarAcc}{a character string. The accession of a categorical vairable.}

\item{...}{There are optional arguments.}

\item{phsAcc}{a character string. (optional) The study accession to which the input variable belongs. This is varCodeVarDF is not provided.}

\item{varCodeValDF}{a data frame. (optional) The variable code value table.}
}
\value{
a named character list. The list of parsed id strings including 'phsAcc' for study acession, 'phsAccNoVer' for the study accession with the v# (version number) removed. 'phsAccId' for study id, and 'phsAccVer' for study version.
}
\description{
The method parses id strings from a dbGaP study accession.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,catVarAcc = character}: A method of class Commons
}}

\keyword{internal}
