% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{checkInputPath}
\alias{checkInputPath}
\alias{checkInputPath,Commons,character-method}
\title{(internal) Validate input directory or file path}
\usage{
checkInputPath(object, path)

\S4method{checkInputPath}{Commons,character}(object, path)
}
\arguments{
\item{object}{Commons class object.}

\item{path}{a character string. A directory or file path.}
}
\value{
a character string. Normalized input path.
}
\description{
The method checks the format of an input directory or file path to see if it is in a valid format. Normalized path is returned.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,path = character}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
checkInputPath(c, path = '/home/foo/temp')
}
}
\keyword{internal}
