% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_internal.R
\docType{methods}
\name{checkPhvAccList}
\alias{checkPhvAccList}
\alias{checkPhvAccList,Study,character-method}
\title{(internal) Validate variable accessions}
\usage{
checkPhvAccList(object, phvAccList)

\S4method{checkPhvAccList}{Study,character}(object, phvAccList)
}
\arguments{
\item{object}{Study class object.}

\item{phvAccList}{a character vector. A list of the dbGaP variable accessions.}
}
\value{
a character vector. A validated list of the variable accessions.
}
\description{
The method validates a given list of the dbGaP variable accessions to make sure each of them is in a right format and belongs to the class study. A new list that have invalid accessions removed is returned.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Study,phvAccList = character}: A method of class Study
}}

\examples{
\dontrun{

s <- Study(phsAcc = 'phs000001.v3.p1')
accList = c('phv00054119.v1.p1.c2', 'phv00053735.v2', 'phv00053732.v2')
checkPhvAccList(s, phvAccList = accList)
}
}
\keyword{internal}
