% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{cleanObjAcc}
\alias{cleanObjAcc}
\alias{cleanObjAcc,Commons-method}
\title{(internal) Validate the format of dbGaP object}
\usage{
cleanObjAcc(object, acc, type)

\S4method{cleanObjAcc}{Commons}(object, acc, type)
}
\arguments{
\item{object}{Commons class object.}

\item{acc}{a character string. The accession of a dbGaP object (study, dataset, or variable object) with or without the p# (participant-set number). The accession returned from the function has the p# removed if it exists in the input string.}

\item{type}{a character string. The dbGaP objec type. The possible values are 'phs' for study, 'pht' for dataset, and 'phv' for variable.}
}
\value{
a character string. The validated accession if the input is valid; An empty string if the input is invalid. The p# is removed from the returned accession if it exists in the input string.
}
\description{
The method checks the format of a dbGaP object accession, including the study, dataset, and variable accessions.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: 
}}

\examples{
\dontrun{

c <- Commons()
cleanObjAcc(c, acc = 'pht003718.v1.p1', type = 'pht')
}
}
\keyword{internal}
