% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commons.R
\docType{class}
\name{Commons-class}
\alias{Commons-class}
\alias{Commons}
\title{Base class of dbgapr package}
\description{
The class contains the functions shared by all functions of the package.
}
\examples{
\dontrun{

c <- Commons()
}
}
