% Generated by hand 
\docType{methods}
\name{commons_internal}
\alias{commons_internal}
\title{Method container}
\description{
    A container of the internal methods of class Commons. 
}
\keyword{internal}

\section{Methods (internal, class Commons)}{
\itemize{
    \item \code{\link{createPrjDir}}
    \item \code{\link{writeLog}}
    \item \code{\link{copyUserFile}}
    \item \code{\link{writeFileInfoJson}}
    \item \code{\link{composeFileInfoJson}}
    \item \code{\link{parsePhenoFileMeta}}
    \item \code{\link{parsePhtHeaderFromFile}}
    \item \code{\link{manageArchive}}
    \item \code{\link{parseDataDicByStudy}}
    \item \code{\link{dataDicXmlParser}}
    \item \code{\link{mergeDatasetConsentByStudy}}
    \item \code{\link{viewAllStudyInfo}}
    \item \code{\link{viewAllDatasetInfo}}
    \item \code{\link{getDataDicByStudy}}
    \item \code{\link{displayTextFile}}
    \item \code{\link{displayTable}}
    \item \code{\link{getVarNameByPhvAcc}}
    \item \code{\link{stripColNameAcc}}
    \item \code{\link{convertEnumVarColName}}
    \item \code{\link{saveGapPlot}}
    \item \code{\link{isStudyAcc}}
    \item \code{\link{parseIdsFromPhtAcc}}
    \item \code{\link{parseIdsFromStAcc}}
    \item \code{\link{cleanObjAcc}}
    \item \code{\link{checkInputPath}}
    \item \code{\link{filterBySubjIds}}
    \item \code{\link{detectVariableDataType}}
    \item \code{\link{getExtData}}
    \item \code{\link{parseDataDic}}
    \item \code{\link{buildVarCodeValCombo}}
}}