% Generated by hand
\docType{methods}
\name{commons_public}
\alias{commons_public}
\title{Method container}
\description{
    A container function that hosts the public methods of class Commons.
}
\keyword{internal}

\section{Methods (public, class Commons)}{
\itemize{
    \item \code{\link{prjConfig}}
    \item \code{\link{getPrjDir}}
    \item \code{\link{prepareData}}
    \item \code{\link{searchCopyPhenoFiles}}
    \item \code{\link{ftpDownload}}
    \item \code{\link{recordPrjFileInfo}}
    \item \code{\link{mergeDatasetConsent}}
    \item \code{\link{accInfo}}
    \item \code{\link{getAllStudyInfo}}
    \item \code{\link{getAllDatasetInfo}}
}}
