% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{composeFileInfoJson}
\alias{composeFileInfoJson}
\alias{composeFileInfoJson,Commons-method}
\title{(internal) Construct a data-frame of the phenotype file info attributes}
\usage{
composeFileInfoJson(object, fileInfoSet)

\S4method{composeFileInfoJson}{Commons}(object, fileInfoSet)
}
\arguments{
\item{object}{Commons class object.}

\item{fileInfoSet}{a character list. The phenotype file info attribute sets.}
}
\value{
a data frame.  Phenotype file info attributes.
}
\description{
The method takes a set of phenotype file info attributes and converts it to a dataframe. It is called through the \code{\link{writeFileInfoJson}} function.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: of class Commons
}}

\examples{
\dontrun{

c <- Commons()
composeFileInfoJson(c, fileInfoSet)
}
}
\keyword{internal}
