% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{convertEnumVarColName}
\alias{convertEnumVarColName}
\alias{convertEnumVarColName,Commons,data.frame-method}
\title{(internal) Convert enumerated integer type variable value}
\usage{
convertEnumVarColName(object, varDataDF, varInfoDF, varCodeValDF)

\S4method{convertEnumVarColName}{Commons,data.frame}(object, varDataDF,
  varInfoDF, varCodeValDF)
}
\arguments{
\item{object}{Commons class object}

\item{varDataDF}{a data frames. The single variable row of the data table.}

\item{varInfoDF}{a data frames. The single variable row of the variable meta-data table.}

\item{varCodeValDF}{a data frames. The code and value table of the categorical vairables.}
}
\description{
The enumerated integer type variable values need to be converted to the respective string type values before plotting. The plotting program treats it as numeric values otherwise. This method takes the data frames of the data and meta-data of an enumerated integer type variable, and converts the variable values to the respective string type values based on the code meanings.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,varDataDF = data.frame}: 
}}

\examples{
\dontrun{

object <- Commons()
convertEnumVarColName(c, varDataDF = catVarDataDF, varInfoDF = catVarInfoDF) 
}
}
\keyword{internal}
