% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{copyUserFile}
\alias{copyUserFile}
\alias{copyUserFile,Commons,character-method}
\title{(internal) Copy decrypted dbGaP files to user project directory}
\usage{
copyUserFile(object, userFile, ...)

\S4method{copyUserFile}{Commons,character}(object, userFile = "", ...,
  copyOrMove = "copy", overwrite = F)
}
\arguments{
\item{object}{Commons class object}

\item{userFile}{a character string. The path to decrypted dbGaP phenotype file.}

\item{...}{There are optional arguments.}

\item{copyOrMove}{a character string. (optional) If the input string is 'copy' (default), copies the file from a user directory to the project directory. Movis the file if the input string is 'move'.}

\item{overwrite}{a logical value. (optional) If TRUE, overwrites the file if it already exists in the prject directory; Skip copy if FALSE (default).}
}
\value{
(invisible) a character list. The meta-info of copied files.
}
\description{
The method copies or moves the decrypted dbgap phenotype files to the project directory. The copied files are organized by study accessions. It is called by the \code{\link{searchCopyPhenoFiles}} function.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,userFile = character}: of class Commons
}}

\examples{
\dontrun{

c <- Commons()
file = '/home/foo/data/phs000429.v1.pht002481.v1.p1.c1.areds_data_final_11.EDO.txt.gz' 
copyUserFile(c, userFile = file, overwrite = TRUE) 
}
}
\keyword{internal}
