% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{createPrjDir}
\alias{createPrjDir}
\alias{createPrjDir,Commons,character-method}
\title{(internal) Create user project directory}
\usage{
createPrjDir(object, prjDir)

\S4method{createPrjDir}{Commons,character}(object, prjDir)
}
\arguments{
\item{object}{Commons class object}

\item{prjDir}{user project directory}
}
\value{
TRUE if the project directory is created and FALSE if not.
}
\description{
The method creates a project directory and updates the project config file. The project directory is where all the data and metadata files are located. The method should be called through the \code{\link{prjConfig}} function.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,prjDir = character}: of class Commons
}}

\examples{
\dontrun{

c <- Commons()
createPrjDir(c, prjDir = "/home/user/dbgapr_project")
# or
createPrjDir(c, prjDir ="C:\\\\Users\\\\user\\\\dbgapr_project")       # for Windows
}
}
\keyword{internal}
