% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{dataDicXmlParser}
\alias{dataDicXmlParser}
\alias{dataDicXmlParser,Commons,character-method}
\title{(internal, deprecated) XML parser of the data dictionary file}
\usage{
dataDicXmlParser(object, xmlFileName, xmlFilePath, phsAcc, ...)

\S4method{dataDicXmlParser}{Commons,character}(object, xmlFileName, xmlFilePath,
  phsAcc, ..., fileInfoDF = data.frame(), computeType = FALSE)
}
\arguments{
\item{object}{Commons class object.}

\item{xmlFileName}{a character string. The name of the data dictionary XML file.}

\item{xmlFilePath}{a character string. The path to the directory where the data dictionary XML file is located.}

\item{phsAcc}{a character string. The dbGaP study accession.}

\item{...}{There are optional arguments.}

\item{fileInfoDF}{a data frame. (optional) The copied file info read from the pht_file_info_indiv.json under the ncbi/dbgapr_conf directory. This is needed when computeType is TRUE.}

\item{computeType}{logical value. (optional) If TRUE, compute data type based on data value and include it in the meta-info file. If FALSE, not compute data type. Note: This step is computationally expensive. Do not set TRUE if it is a large study that have a large number of datasets.}
}
\value{
a character string. The path to the resulting rds file of the parsed data dictionary.
}
\description{
The method parses the field values of a given data dictionary XML file. It is called through the function "parseDataDicByStudy".
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,xmlFileName = character}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
filename = 'phs000001.v3.pht000001.v2.genspecphenotype.data_dict.xml'
filepath = '/home/foo/my_project/gapwork/data/phs000001/phs000001.v3/data_dict'
dataDicXmlParser(c, xmlFileName=filename, xmlFilePath=filepath, phsAcc='phs000001.v3.p1')
}
}
\keyword{internal}
