% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{displayTable}
\alias{displayTable}
\alias{displayTable,Commons-method}
\title{(internal) Display text file content}
\usage{
displayTable(object, ...)

\S4method{displayTable}{Commons}(object, ..., file = "",
  data = data.frame())
}
\arguments{
\item{object}{Commons class object.}

\item{...}{There are optional arguments.}

\item{file}{a character string. (optional) The path to the data file to display in rds, csv, or json format.}

\item{data}{a data frame. (optional) The data to display.}
}
\value{
a data frame. The data for display.
}
\description{
The method displays the input data frame as a data table through a platform specific table viewer. The data frame is either input directly as a function argument or through an input data file of rds, csv, or json format.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
df <- readRDS(data_dic.rds) 
displayTable(c, data = df)
displayTable(c, file = 'data_dic.json')
displayTable(c, file = 'data_dic.csv')
displayTable(c, file = 'data_dic.rds')
}
}
\keyword{internal}
