% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{displayTextFile}
\alias{displayTextFile}
\alias{displayTextFile,Commons,character-method}
\title{(Internal) Display text file content}
\usage{
displayTextFile(object, file, ...)

\S4method{displayTextFile}{Commons,character}(object, file, ..., editor = "")
}
\arguments{
\item{object}{Commons class object.}

\item{file}{a character string. The path to the text file to display.}

\item{...}{There are optional arguments.}

\item{editor}{a character string. (optional) The name of your favorite plain text editor. It should be executable from a command-line prompt of the respective platform. For example, notepad (Windows), vim, emacs (Unix), gedit (Ubuntu), nedit (CentOS), etc.}
}
\description{
The method displays the content of a given test file.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,file = character}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
file = 'phs000001.v3_pht000371.v2_data_dic.json'
displayTextFile(c, file = file)
}
}
\keyword{internal}
