% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{filterBySubjIds}
\alias{filterBySubjIds}
\alias{filterBySubjIds,Commons,data.frame-method}
\title{(internal) Filter dataset data by subject ids}
\usage{
filterBySubjIds(object, varDF, ...)

\S4method{filterBySubjIds}{Commons,data.frame}(object, varDF, ...,
  subjIdsOrFile)
}
\arguments{
\item{object}{Commons class object.}

\item{varDF}{a data frame. The dataset data.}

\item{...}{There are optional arguments.}

\item{subjIdsOrFile}{a character vector or a character string. (optional) This argument is either list of subject ids (dbGaP_Subject_ID) shared among the input variables or the path to a file that contains a list of of the subject ids shared among the input variables. The file is a plain text file with one dbGaP_Subject_ID per line.}
}
\value{
a data frame. Filtered dataset data.
}
\description{
The method filters out the rows of a dataset by a given list of subject ids. The subject ids can be provided as a list or as a text file.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,varDF = data.frame}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
filterBySubjIds(c, varDF = varDF, subjIdsOrFile = c("219", "220", "223"))
}
}
\keyword{internal}
