% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_public.R
\docType{methods}
\name{ftpDownload}
\alias{ftpDownload}
\alias{ftpDownload,Commons-method}
\title{Download supplemental data files from dbGaP ftp}
\usage{
ftpDownload(object, ...)

\S4method{ftpDownload}{Commons}(object, ..., phsAcc = "")
}
\arguments{
\item{object}{Commons class object}

\item{...}{There are optional arguments.}

\item{phsAcc}{a character string. The dbGaP study accession. Only the files of the given study are downloaded when a study accession is provided.}
}
\value{
a logical value. If TRUE, all files are downloaded OK. If FALSE, there is at least one failed download.
}
\description{
The method downloads the supplemental metadata files from the dbGaP ftp site. It is called by the \code{\link{prepareData}} function.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
ftpDownload(c)
# or
ftpDownload(c, phsAcc = 'phs000001.v3.p1')
}
}
