% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_public.R
\docType{methods}
\name{getAllDatasetInfo}
\alias{getAllDatasetInfo}
\alias{getAllDatasetInfo,Commons-method}
\title{Get meta-info of all datasets}
\usage{
getAllDatasetInfo(object, ...)

\S4method{getAllDatasetInfo}{Commons}(object, ..., showAs = "", editor = "")
}
\arguments{
\item{object}{Commons class object.}

\item{...}{There are optional arguments.}

\item{showAs}{a character string. (optional) When the value is 'table', displays the data as a table through a platform specific table viewer; When it is 'json', displays the json text through a plain text editor; When it is 'text', displays in a brief left-justified text format.}

\item{editor}{a character string. (optional) The name of your favorite plain text editor. It should be executable from a command-line prompt of the respective platform. For example, notepad (Windows), vim, emacs (Unix), gedit (Ubuntu), nedit (CentOS), etc.}
}
\value{
a data frame. The meta-info of all datasets available under the project directory.
}
\description{
The method returns meta-info of all datasets of the studies that have data files available under the project directory.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
getAllDatasetInfo(c)
getAllDatasetInfo(c, showAs = 'json', editor='notepad')
getAllDatasetInfo(c, showAs = 'text', editor='gedit')
}
}
