% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_public.R
\docType{methods}
\name{getAllStudyInfo}
\alias{getAllStudyInfo}
\alias{getAllStudyInfo,Commons-method}
\title{Get meta-info of all studies}
\usage{
getAllStudyInfo(object, ...)

\S4method{getAllStudyInfo}{Commons}(object, ..., showAs = "", editor = "")
}
\arguments{
\item{object}{Commons class object.}

\item{...}{There are optional arguments.}

\item{showAs}{a character string. (optional) When the value is 'table', displays the data as a table through a platform specific table viewer; When it is 'json', displays the json text through a plain text editor; When it is 'text', displays in a brief left-justified text format.}

\item{editor}{a character string. (optional) The name of your favorite plain text editor. It should be executable from a command-line prompt of the respective platform. For example, notepad (Windows), vim, emacs (Unix), gedit (Ubuntu), nedit (CentOS), etc.}
}
\value{
a data frame. The meta-info of all studies. For the table row of a non-cohort study or a parent study in a chort, the root_study and this_study are identical. For the table row of a child- (or sub-) study in a cohort, the this_study column represents the child study, and the root_study column represents the parent study.
}
\description{
The method returns meta-info of all studies available in the project directory. In case of cohort studies, both parent and child studies (also called root- and sub-studies) are included.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
getAllStudyInfo(c)
# or
getAllStudyInfo(c, showAs = 'json', editor = 'notepad')
# or
getAllStudyInfo(c, showAs = 'text', editor = 'gedit')
}
}
