% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{getAvailPhtVer}
\alias{getAvailPhtVer}
\alias{getAvailPhtVer,Commons,character-method}
\title{(internal) Get available dataset accession by variable accession}
\usage{
getAvailPhtVer(object, randPhtAcc, ...)

\S4method{getAvailPhtVer}{Commons,character}(object, randPhtAcc, ...,
  phsAcc = NA, studyPhtInfoDF = data.frame(), dataStudyOnly = TRUE)
}
\arguments{
\item{object}{Commons class object.}

\item{randPhtAcc}{a character string. One of the randomly chosen dataset accessions returned from queyring study_variable_info table by a variable accession.}

\item{...}{There are optional arguments.}

\item{phsAcc}{a character string. A study accession. It is required when studyPhtInfoDF is not provided.}

\item{studyPhtInfoDF}{a data frame. The dataset info table of the study.}

\item{dataStudyOnly}{a logical value. When TRUE (default), only downloads the dataset and variable metadata of the stdudies that have data files in the project directory.  When FALSE, downloads the dataset and variable metadata of all dbGaP released studies, regardless the actual phenotype data files of the studies are downloaded or not.}
}
\value{
a character string. The dataset accession that have a match with available studies.
}
\description{
When querying study_variable_info table with a variable accession, multiple dataset versions could be returned. This methods further queris the study_dataset table given any of the dataset accessions from the early query, and returns the correct dataset accession that matches the given study accession.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,randPhtAcc = character}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
getAvailPhtVer(c, randPhtAcc='pht000009.v1', phsAcc='phs000007.v29')
}
}
\keyword{internal}
