% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{getDataDicByStudy}
\alias{getDataDicByStudy}
\alias{getDataDicByStudy,Commons,character-method}
\title{(internal) Get study variable meta-info (data dictionary)}
\usage{
getDataDicByStudy(object, phsAcc)

\S4method{getDataDicByStudy}{Commons,character}(object, phsAcc)
}
\arguments{
\item{object}{Commons class object.}

\item{phsAcc}{a character string. The dbGaP study accession.}
}
\value{
a data frame. The dataset meta-info of the study.
}
\description{
The method returns the meta-info of all variables of the given study.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,phsAcc = character}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
getDataDicByStudy(c, phsAcc = 'phs000001.v3.p1')
}
}
\keyword{internal}
