% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_internal.R
\docType{methods}
\name{getDatasetDataByPhtAcc}
\alias{getDatasetDataByPhtAcc}
\alias{getDatasetDataByPhtAcc,Study,character-method}
\title{(internal) Get data by dataset accession}
\usage{
getDatasetDataByPhtAcc(object, phtAcc, ...)

\S4method{getDatasetDataByPhtAcc}{Study,character}(object, phtAcc, ...,
  dbgapIdsOrFile = NULL, colNameWithAcc = FALSE)
}
\arguments{
\item{object}{Study class object.}

\item{phtAcc}{a character string. The dbGaP phenotype dataset accession.}

\item{...}{There are optional arguments.}

\item{dbgapIdsOrFile}{a character vector or a character string. (optional) This argument can be either a vector of ID list or a path to a file that contains a list of IDs. The IDs can be dbGaP_Subject_ID or dbGaP_Sample_ID denpending on type of the data. When the list of IDs is provided by a file, it should be  a plain text file with one ID per line.}

\item{colNameWithAcc}{logical value. (optional). If TRUE, includes the variable accessions in the column names (e.g. AGEPHOT_phv00000027.v2); If FALSE, not include (e.g. AGEPHOT).}
}
\value{
a data frame. Data of the dataset.
}
\description{
The method returns the data of the given dataset. For large dataset, it may take quite a while to retrieve the data. The data of both subject and sample level variables can be returned through this method.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Study,phtAcc = character}: A method of class Study
}}

\examples{
\dontrun{

s <- Study(phsAcc = 'phs000001.v3.p1')
getDatasetDataByPhtAcc(s, phtAcc = 'pht000370.v2.p1')
# or
getDatasetDataByPhtAcc(s, phtAcc = 'pht000370.v2.p1', colNameWithAcc = TRUE)
}
}
\keyword{internal}
