% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{getDatasetMetaByStudy}
\alias{getDatasetMetaByStudy}
\alias{getDatasetMetaByStudy,Commons,character-method}
\title{(internal) Get study variable meta-info (data dictionary)}
\usage{
getDatasetMetaByStudy(object, phsAcc)

\S4method{getDatasetMetaByStudy}{Commons,character}(object, phsAcc)
}
\arguments{
\item{object}{Commons class object.}

\item{phsAcc}{a character string. The dbGaP study accession.}
}
\value{
a data frame. The dataset meta-info of the study.
}
\description{
The method returns the meta-info of all datasets of the given study.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,phsAcc = character}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
getDatasetMetaByStudy(c, phsAcc = 'phs000001.v3.p1')
}
}
\keyword{internal}
