% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_public.R
\docType{methods}
\name{getIdInfo}
\alias{getIdInfo}
\alias{getIdInfo,Study-method}
\title{Get subject, sample, or pedigree information.}
\usage{
getIdInfo(object, ...)

\S4method{getIdInfo}{Study}(object, infoType, ..., colNameWithAcc = FALSE)
}
\arguments{
\item{object}{Study class object.}

\item{...}{There are optional arguments.}

\item{infoType}{a charcter string. The type of information that is either 'subject', 'sample', or 'pedigree'.}

\item{colNameWithAcc}{a logical value. (optional) If TRUE, the variable column name is concatenated with the respective variable accession (e.g. AGEPHOT_phv00000027.v2); If FALSE (default), keep the original column name unchanged (e.g. AGEPHOT).}
}
\value{
a data frame. The subject, sample, or pedigree info data table.
}
\description{
The method retrieves either subject, sample, or pedigree information of a given study.
}
\section{Methods (by class)}{
\itemize{
\item \code{Study}: of class Study
}}

\examples{
\dontrun{

s <- Study(phsAcc = 'phs000001.v3.p1')
getIdInfo(s)
}
}
