% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{getMetaByObjAcc}
\alias{getMetaByObjAcc}
\alias{getMetaByObjAcc,Commons-method}
\title{Get metadta of a dbGaP object}
\usage{
getMetaByObjAcc(object, acc, type)

\S4method{getMetaByObjAcc}{Commons}(object, acc, type)
}
\arguments{
\item{object}{Commons class object.}

\item{acc}{a character string. The dbGaP dataset, or variable accession.}

\item{a}{character string. The type of the input accession that can be either 'dataseet' or 'variable'.}
}
\value{
a data frame. The metadata of the input object.
}
\description{
The method returns the metadat a given dataset or variable accession.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: returns dataset accession
}}

\examples{
\dontrun{

c <- Commons()
df <- getMetaByObjAcc(c, acc = 'pht000824.v5', type = 'dataset')
# or
df <- getMetaByObjAcc(c, acc = 'phv00000090.v2', type = 'variable')
}
}
\keyword{internal}
