% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_public.R
\docType{methods}
\name{getPrjDir}
\alias{getPrjDir}
\alias{getPrjDir,Commons-method}
\title{Get user project directory}
\usage{
getPrjDir(object, ...)

\S4method{getPrjDir}{Commons}(object, ..., showErr = as.logical(F))
}
\arguments{
\item{object}{Commons class object.}

\item{...}{There are optional arguments.}

\item{showErr}{a logical value. (optional) If TRUE, display the error message; If FALSE (default), not display.}
}
\value{
a character list. Current user project directory and the sub-directories.
}
\description{
The method looks for the current project directory information and associates the directory and its sub-directories to the class object. It returns a list of paths to the current project directory and the sub-directories. The method is called every time when a class object is initialized, which makes the current project directory and the sub-directory information readily accessible through the class object.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: of Class Commons
}}

\examples{
\dontrun{

c <- Commons()
getPrjDir(c, showErr = TRUE)
}
}
