% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_public.R
\docType{methods}
\name{getStudyDatasetInfo}
\alias{getStudyDatasetInfo}
\alias{getStudyDatasetInfo,Study-method}
\title{Get meta-info of study datasets}
\usage{
getStudyDatasetInfo(object, ...)

\S4method{getStudyDatasetInfo}{Study}(object, ..., phtAcc = "",
  dataStudyOnly = TRUE, showAs = "", editor = "")
}
\arguments{
\item{object}{Study class object.}

\item{...}{There are optional arguments.}

\item{phtAcc}{a character string. The dbGaP dataset accession.}

\item{dataStudyOnly}{a logical value. When TRUE (default), only downloads the dataset and variable metadata of the stdudies that have data files in the project directory.  When FALSE, downloads the dataset and variable metadata of all dbGaP released studies, regardless the actual phenotype data files of the studies are downloaded or not.}

\item{showAs}{a character string. (optional) When the value is 'table', displays the data as a table through a platform specific table viewer; When it is 'json', displays the json text through a plain text editor; When it is 'text', displays in a brief left-justified text format.}

\item{editor}{a character string. (optional) The name of your favorite plain text editor. It should be executable from a command-line prompt of the respective platform. For example, notepad (Windows), vim, emacs (Unix), gedit (Ubuntu), nedit (CentOS), etc.}
}
\value{
a data frame. The meta-info of input variables.
}
\description{
The method returns the dataset info of the study. When dataset accession is not providedd, it returns the meta-info of all datasets of the class study. Only the spcified dataset info is returned when the dataset accession is provided.
}
\section{Methods (by class)}{
\itemize{
\item \code{Study}: of class Study
}}

\examples{
\dontrun{

s <- Study(phsAcc = 'phs000001.v3.p1')
# all study datasets
getStudyDatasetInfo(s)
# or a specific dataset
getStudyDatasetInfo(s, phtAcc = 'pht000370.v2.p1')
# or a dataset of a study that has no data file under project directory
s2 <- Study(phsAcc = 'phs001255.v1')
getStudyDatasetInfo(s2, phtAcc = 'pht005990.v1', dataStudyOnly=F)
}
}
