% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_internal.R
\docType{methods}
\name{getStudyVariableInfoByDataType}
\alias{getStudyVariableInfoByDataType}
\alias{getStudyVariableInfoByDataType,Study,character-method}
\title{(internal) Get meta-info of study numeric or categorical variables}
\usage{
getStudyVariableInfoByDataType(object, dataType, ...)

\S4method{getStudyVariableInfoByDataType}{Study,character}(object, dataType,
  ..., dataDicDF = data.frame(), dataStudyOnly = TRUE)
}
\arguments{
\item{object}{Study class object.}

\item{dataType}{a character string. The variable datatype. The possible value is either 'num' (for numeric variable) or 'cat' (for categorical variable).}

\item{...}{There are optional arguments.}

\item{dataDicDF}{a data frame. (optional) Dataset data dictionary data.}

\item{dataStudyOnly}{a logical value. When TRUE (default), only downloads the dataset and variable metadata of the stdudies that have data files in the project directory.  When FALSE, downloads the dataset and variable metadata of all dbGaP released studies, regardless the actual phenotype data files of the studies are downloaded or not.}
}
\value{
a data frame. The variable meta-info.
}
\description{
The method returns the variable meta-info of either numeric or categorical variables of the study.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Study,dataType = character}: A method of class Study
}}

\examples{
\dontrun{

getStudyVariableInfoByDataType(s, dataType = 'num', dataDicDF = dataDicDF)
getStudyVariableInfoByDataType(s, dataType = 'cat', dataDicDF = dataDicDF)
}
}
\keyword{internal}
