% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_internal.R
\docType{methods}
\name{getVariableDataByPhvAcc}
\alias{getVariableDataByPhvAcc}
\alias{getVariableDataByPhvAcc,Study,character-method}
\title{(internal) Get variable data by accession}
\usage{
getVariableDataByPhvAcc(object, phvAccList, ...)

\S4method{getVariableDataByPhvAcc}{Study,character}(object, phvAccList, ...,
  dataStudyOnly = TRUE, cleanNumVal = FALSE, emptyToNa = FALSE,
  colNameWithAcc = FALSE, checkList = T)
}
\arguments{
\item{object}{Study class object.}

\item{phvAccList}{a character vector. The dbGaP variable accessions.}

\item{...}{There are optional arguments.}

\item{dataStudyOnly}{a logical value. When TRUE (default), only downloads the dataset and variable metadata of the stdudies that have data files in the project directory.  When FALSE, downloads the dataset and variable metadata of all dbGaP released studies, regardless the actual phenotype data files of the studies are downloaded or not.}

\item{cleanNumVal}{a logical value. (optional). If TRUE, from numeric variables, converts non-numeric values to NA. If FALSE (default), not convert.}

\item{emptyToNa}{a logical value. (optional). If TRUE, converts the empty values to NA; If FALSE (default), not convert.}

\item{colNameWithAcc}{a logical value. (optional). If TRUE, includes the variable accessions in the column names (e.g. AGEPHOT_phv00000027.v2); If FALSE, not include (e.g. AGEPHOT).}

\item{checkList}{a logical value. (optional). If TRUE, validate the input variable accessions. Skip the check if FALSE.}
}
\value{
a data frame. Merged data of input variables.
}
\description{
The method returns the data frame of merged data of given dbGaP variable accessions. The variables need to belong to the same study. Only the data of subject level variables can be returned by this function.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Study,phvAccList = character}: returns a dataframe of the merged variable data.
}}

\examples{
\dontrun{

s <- Study(phsAcc = 'phs000001.v3.p1')
var_list <- c('phv00054119.v1.p1.c2', 'phv00054118.v1.p1', 'phv00053733.v2')
getVariableDataByPhvAcc(s, phvAccList = var_list)
}
}
\keyword{internal}
