% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_internal.R
\docType{methods}
\name{getVariableInfoByPhvAcc}
\alias{getVariableInfoByPhvAcc}
\alias{getVariableInfoByPhvAcc,Study,character-method}
\title{(internal) Get variable meta-info by accession}
\usage{
getVariableInfoByPhvAcc(object, phvAccList, ...)

\S4method{getVariableInfoByPhvAcc}{Study,character}(object, phvAccList, ...,
  showTable = F, showBrief = T, validateInput = T)
}
\arguments{
\item{object}{Study class object.}

\item{phvAccList}{a character vector. The dbGaP phenotype variable accessions.}

\item{...}{There are optional arguments.}

\item{showTable}{(optional) a logical value. If TRUE, displays the variable meta-info in a platform specific table viewer; Not display if FALSE (default).}

\item{showBrief}{(optional) a logical value. If TRUE (default), console displays a brief version of the variable info. Not display if FALSE.}

\item{validateInput}{(optional). If TRUE, checks the input variable accessions to make sure they belong to downloaded data under the user project. No check if FALSE.}
}
\value{
a data frame. (invisible) The meta-info of the input variables.
}
\description{
The method returns the variable meta-info as a data frame given the variable accession.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Study,phvAccList = character}: of class Commons
}}

\examples{
\dontrun{

s <- Study(phsAcc = 'phs000001.v3.p1')
accList = c('phv00054119.v1.p1', 'phv00054118.v1.p1', 'phv00000035.v2')
getVariableInfoByPhvAcc(s, phvAccList = accList) 
getVariableInfoByPhvAcc(s, phvAccList = accList, showBrief=T)
}
}
\keyword{internal}
