% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{getVarNameByPhvAcc}
\alias{getVarNameByPhvAcc}
\alias{getVarNameByPhvAcc,Commons,character-method}
\title{(internal) Get variable names given accessions}
\usage{
getVarNameByPhvAcc(object, phvAccList, ...)

\S4method{getVarNameByPhvAcc}{Commons,character}(object, phvAccList, ...,
  colNameWithAcc = F, underscore = F, phsAcc,
  studyDataDicDF = data.frame())
}
\arguments{
\item{object}{Commons class object.}

\item{phvAccList}{a character vector. The dbGaP phenotype variable accessions.}

\item{...}{There are optional arguments.}

\item{colNameWithAcc}{a logical value. (optional). If FALSE (default), outputs variable name only; If TRUE, outputs concatenated variable name and variable.}

\item{underscore}{a logical value. (optional) Used only when colNameWithAcc is TRUE. If TRUE (default), underscore is used between variable name and accession( e.g. LNUCSCORE_phv00000006.v2 ); If FALSE, the parenthesis with a space is used between variable name and accession ( e.g. LNUCSCORE (phv00000006.v2) ).}

\item{phsAcc}{a character string. (optional) The accession of the study where the variables belong. It is required if studyDataDicDF is not provided.}

\item{studyDataDicDF}{a data frame. (optional) The data frame of the study data dictionary (variable meta-data). It is required if phsAcc is not provided.}
}
\value{
a named list of variable names. The variable names of the input variable accession.
}
\description{
The method takes a list of variable accessions and returns a list of variable names. The output variable name can be shown in three different formats: name only ( e.g. LNUCSCORE ), name and accession separated by underscore (e.g. LNUCSCORE_phv00000006.v2), name and accession separated with space and parenthesis (e.g. LNUCSCORE (phv00000006.v2)). There is no validation applied to the input accessions.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,phvAccList = character}: 
}}

\examples{
\dontrun{

c <- Commons()
accList = c('phv00000027.v2', 'phv00053747.v2')
phsAcc = 'phs000001.v3.p1'
getVarNameByPhvAcc(c, phvAccList=accList, phsAcc=phsAcc) 
# or
getVarNameByPhvAcc(c, phvAccList=accList, phsAcc=phsAcc, colNameWithAcc=T)
# or
getVarNameByPhvAcc(c, phvAccList=accList, phsAcc=phsAcc, colNameWithAcc=T, underscore=T)
}
}
\keyword{internal}
