% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commons.R
\docType{methods}
\name{initialize,Commons-method}
\alias{initialize,Commons-method}
\title{Initialize class Commons}
\usage{
\S4method{initialize}{Commons}(.Object, ...)
}
\arguments{
\item{.Object}{The class object.}

\item{...}{There arer optional arguments.}
}
\description{
The method initializes the Commons class, through which the package wide global variables are made avaiable through the class object.
}
\examples{
\dontrun{

c <- Commons()
}
}
