% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\docType{methods}
\name{initialize,Study-method}
\alias{initialize,Study-method}
\title{Initialize Study class}
\usage{
\S4method{initialize}{Study}(.Object, ..., phsAcc = .Object@phsAcc)
}
\arguments{
\item{.Object}{private arugment used in class init}

\item{...}{There are optional arguments.}

\item{phsAcc}{a character string. A dbGaP study accession.}
}
\description{
The method initializes the Study class. Once the class is initialized, the input study accession and all varibles associated with the parent class Commons become available to all classs methods through the class object.
}
\examples{
\dontrun{

s <- new("Study", phsAcc = 'phs000001.v1.p1')
s <- Study(phsAcc = 'phs000001.v1.p1')
}

}
