% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{isStudyAcc}
\alias{isStudyAcc}
\alias{isStudyAcc,Commons,character-method}
\title{(internal) Validate the format of a dbGaP study accession}
\usage{
isStudyAcc(object, phsAcc)

\S4method{isStudyAcc}{Commons,character}(object, phsAcc)
}
\arguments{
\item{object}{Commons class object.}

\item{phsAcc}{a character string. A dbGaP study accession.}
}
\value{
a logical value. TRUE if the format matches a dbGaP accession, FALSE if it doesn't match.
}
\description{
The method checks the format of an input string to see if it is a dbGaP accession. The accessions with and without the p# (participant-set number) both are considered to be correct (e.g. phs000001.v3.p1 or phs000001.v3).
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,phsAcc = character}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
isStudyAcc(c, phsAcc = 'phs000001.v3')
}
}
\keyword{internal}
