% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{manageArchive}
\alias{manageArchive}
\alias{manageArchive,Commons,character-method}
\title{(internal) Logging archive manager}
\usage{
manageArchive(object, workFile, archiveDir, type, ...)

\S4method{manageArchive}{Commons,character}(object, workFile, archiveDir,
  type = "infojson", ..., maxSizeInMb = 1, keptDay = 90, minFileAge = 1)
}
\arguments{
\item{object}{Commons class object.}

\item{workFile}{a character string. The path to the log file, such as project_setup.log.}

\item{archiveDir}{a character string. The path to the archive directory.}

\item{type}{a character string. The logging type of either 'setup' for project_setup.log or 'process' for data_process.log.}

\item{...}{There are optional arguments.}

\item{maxSizeInMb}{an integer. (optional) The maximum size (MB) of a log file. The log file is moved to the archived_log directory after reaching the maxSizeInDb (default 1MB).}

\item{keptDay}{an integer. (optional) The maximum days of an archived log file is kept. The archived log file is deleted after the durationDay (default 90 days).}

\item{minFileAge}{an integer. (optional) The minimum age in days of a file can be archived.}
}
\description{
The method moves expired log files into archive and deletes old archived files based on given conditions.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,workFile = character}: A method of class Commons
}}

\keyword{internal}
