% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_public.R
\docType{methods}
\name{mergeDatasetConsent}
\alias{mergeDatasetConsent}
\alias{mergeDatasetConsent,Commons-method}
\title{Merge dataset data of different consents}
\usage{
mergeDatasetConsent(object, ...)

\S4method{mergeDatasetConsent}{Commons}(object, ..., phsAcc = "",
  overwrite = TRUE)
}
\arguments{
\item{object}{Commons class object.}

\item{...}{There are optional arguments.}

\item{phsAcc}{a character string. The dbGaP study accession.}

\item{overwrite}{a logical value. (optional) If TRUE (default), overwrites existing files. If FALSE, skips the merge for entire study if the merged file directory of the respective study is not empty.}
}
\value{
a data frame. (invisible) The meta-info of merged data files.
}
\description{
The method searches for all dataset files under the project directory and merges the files that have the same dataset accession but belong to different consents.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
mergeDatasetConsent(c)
mergeDatasetConsent(c, phsAcc = "phs000001.v3.p1")
}
}
