% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{mergeDatasetConsentByStudy}
\alias{mergeDatasetConsentByStudy}
\alias{mergeDatasetConsentByStudy,Commons,character-method}
\title{(internal) Merge data of different consent groups in a study}
\usage{
mergeDatasetConsentByStudy(object, phsAcc)

\S4method{mergeDatasetConsentByStudy}{Commons,character}(object, phsAcc)
}
\arguments{
\item{object}{Commons class object.}

\item{phsAcc}{a character string. The dbGaP study accession.}
}
\value{
a data frame. Meta-info of the merged data files.
}
\description{
The method merges the dataset files that have the same dataset accession but different consents into one file.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,phsAcc = character}: returns study accession
}}

\examples{
\dontrun{

c <- Commons()
mergeDatasetConsentByStudy(c, phsAcc = 'phs000001.v3.p1')
}
}
\keyword{internal}
