% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{parseDataDic}
\alias{parseDataDic}
\alias{parseDataDic,Commons-method}
\title{(internal, deprecated) Parse data dictionary XML files of all studies.}
\usage{
parseDataDic(object, ...)

\S4method{parseDataDic}{Commons}(object, ..., phsAcc = "",
  computeType = FALSE)
}
\arguments{
\item{object}{Commons class object.}

\item{...}{There are optional arguments.}

\item{phsAcc}{a character string. The dbGaP study accession.}

\item{computeType}{logical value. (optional) If TRUE, compute data type based on data value and include it in the meta-info file. If FALSE, not compute data type. Note: This step is computationally expensive. Do not use it if it is a large study such as Framingham.}
}
\value{
(invisible) a data frame. The combined data of data dictionary files of all studies.
}
\description{
The method parses the data dictionary XML files of a specific or all studies available under the project directory. The parsed values from one study are merged and saved in files of three different formats (rds, csv, and json). Before calling the function, the data dictionary files of the study should be made available in the data_dic sub-directory under the project directory. Checkout the \code{\link{searchCopyPhenoFiles}} function for how to move the files to the directory if not. The method is no longer needed since the data dictionary XML files are replaced by the csv format of the data dictionary downloadable from the dbGaP ftp site.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
parseDataDic(c) 
# or
parseDataDic(c, phsAcc = 'phs000001.v3.p1') 
}
}
\keyword{internal}
