% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{parseDataDicByStudy}
\alias{parseDataDicByStudy}
\alias{parseDataDicByStudy,Commons,character-method}
\title{(internal, deprecated) Parse data dictionary XML files of a specified study}
\usage{
parseDataDicByStudy(object, phsAcc, ...)

\S4method{parseDataDicByStudy}{Commons,character}(object, phsAcc, ...,
  computeType = TRUE)
}
\arguments{
\item{object}{Commons class object.}

\item{phsAcc}{a character string. The dbGaP study accession.}

\item{...}{There are optional arguments.}

\item{computeType}{logical value. (optional) If TRUE, compute data type based on data value and include it in the meta-info file. If FALSE, not compute data type. Note: This step is computationally expensive. Do not use it if it is a large study such as Framingham.}
}
\value{
(invisible) a data frame. The merged content of the parsed data dictionary files.
}
\description{
The method parses out the values from the data dictionary XML files of the given study. The parsed values are merges into a single data frame, which is further saved to files of various formats (rds, csv, and json). Before calling the function, the data dictionary files of the study should be made available in the data_dic sub-directory under the project directory. Checkout the \code{\link{searchCopyPhenoFiles}} function for how to move the files to the directory if not. In most of cases, this function is called through \code{\link{parseDataDic}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,phsAcc = character}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
parseDataDicByStudy(c, phsAcc = 'phs000001.v3.p1') 
parseDataDicByStudy(c, phsAcc = 'phs000007.v27') 
}
}
\keyword{internal}
