% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{parseIdsFromPhtAcc}
\alias{parseIdsFromPhtAcc}
\alias{parseIdsFromPhtAcc,Commons-method}
\title{Parse ids from dbGaP dataset accession}
\usage{
parseIdsFromPhtAcc(object, phtAcc)

\S4method{parseIdsFromPhtAcc}{Commons}(object, phtAcc)
}
\arguments{
\item{object}{Commons class object.}

\item{phtAcc}{a character string. The dbGaP dataset accession with or without the p# (participant-set number), e.g. pht003718.v1.p1 or pht003718.v1.}
}
\value{
a named character list. The list of parsed id strings including 'phtAcc' for dataset acession, 'phtAccNoVer' for the dataset accession with the v# (version number) removed, 'phtAccId' for dataset id, and 'phtAccVer' for dataset version.
}
\description{
The method parses out id strings from a dbGaP dataset accession.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
parseIdsFromPhtAcc(c, phtAcc = 'pht003718.v1.p1')
}
}
\keyword{internal}
