% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{parseIdsFromPhvAcc}
\alias{parseIdsFromPhvAcc}
\alias{parseIdsFromPhvAcc,Commons-method}
\title{Parse ids from dbGaP variable accession}
\usage{
parseIdsFromPhvAcc(object, phvAcc)

\S4method{parseIdsFromPhvAcc}{Commons}(object, phvAcc)
}
\arguments{
\item{object}{Commons class object.}

\item{phvAcc}{a character string. The dbGaP variable accession with or without the p# (participant-set number), e.g. phv00273621.v1.}
}
\value{
a named character list. The list of parsed id strings including 'phvAcc' for dataset acession, 'phvAccNoVer' for the dataset accession with the v# (version number) removed, 'phvAccId' for dataset id, and 'phvAccVer' for variable version.
}
\description{
The method parses out id strings from a dbGaP variable accession.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
phvAccIds <- parseIdsFromPhvAcc(c, phvAcc = 'phv00273621.v1.p1')
}
}
\keyword{internal}
