% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{parseIdsFromStAcc}
\alias{parseIdsFromStAcc}
\alias{parseIdsFromStAcc,Commons-method}
\title{Parse ids from dbGaP study accession}
\usage{
parseIdsFromStAcc(object, phsAcc)

\S4method{parseIdsFromStAcc}{Commons}(object, phsAcc)
}
\arguments{
\item{object}{Commons class object.}

\item{phsAcc}{a character string. A dbGaP study accession with or without the p# (participant-set number), e.g. phs000001.v3.p1 or phs000001.v3.}
}
\value{
a named character list. The list of parsed id strings including 'phsAcc' for study accession, 'phsAccNoVer' for the study accession with the v# (version number) removed. 'phsAccId' for study id, and 'phsAccVer' for study version.
}
\description{
The method parses out id strings from a dbGaP study accession.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons() 
parseIdsFromStAcc(c, phsAcc = 'phs000001.v3.p1')
}
}
\keyword{internal}
