% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{parsePhtHeaderFromFile}
\alias{parsePhtHeaderFromFile}
\alias{parsePhtHeaderFromFile,Commons,character-method}
\title{(internal) Parse phenotype data file header}
\usage{
parsePhtHeaderFromFile(object, phtFile, phtFileType)

\S4method{parsePhtHeaderFromFile}{Commons,character}(object, phtFile,
  phtFileType)
}
\arguments{
\item{object}{Commons class object.}

\item{phtFile}{Path to a phenotype file.}

\item{phtFileType}{Type of the phenotype file with possible values of Subject, Sample, Pedigree, or Other.}
}
\value{
a list of the file attribute sets, including the sets of variable accessions, column names, variable accession and name pairs, variable accession and name combo strings.
}
\description{
The method parses the header lines of a given phenotype data file and returns a list of file attribute sets. It is called through \code{\link{parsePhenoFileMeta}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,phtFile = character}: A method of class Commons
}}

\keyword{internal}
