% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_public.R
\docType{methods}
\name{prepareData}
\alias{prepareData}
\alias{prepareData,Commons,character-method}
\title{Copy and preprocess data files}
\usage{
prepareData(object, userDataDir, ...)

\S4method{prepareData}{Commons,character}(object, userDataDir, ...,
  phsAcc = "", dataStudyOnly = TRUE)
}
\arguments{
\item{object}{Commons class object.}

\item{userDataDir}{a character string. The path to the top level directory of decrypted dbGaP phenotype data files.}

\item{...}{There are optional arguments.}

\item{phsAcc}{a character string. The dbGaP study accession.}

\item{dataStudyOnly}{a logical value. When TRUE (default), only downloads the dataset and variable metadata of the stdudies that have data files in the project directory.  When FALSE, downloads the dataset and variable metadata of all dbGaP released studies, regardless the actual phenotype data files of the studies are downloaded or not.}
}
\value{
a data frame. (invisible) The meta-info of merged data files.
}
\description{
The method wraps several function calls related to the data preparation into one call. It copies user data files to the project directory ( \code{\link{searchCopyPhenoFiles}} ), downloads supplemental metadata files from the dbGaP ftp site ( \code{\link{ftpDownload}} ), extracts and saves the file meta-info ( \code{\link{recordPrjFileInfo}} ), and finally merges the data files of different consents ( \code{\link{mergeDatasetConsent}} ). If any of above steps hasn't gone through successfully (often due to temporary glitches), rerun the failed and rest of steps individually to make sure all steps are finished before going forward.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,userDataDir = character}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
prepareData(c, userDataDir = '/home/user/data')
# or
prepareData(c, userDataDir = '/home/user/data', phsAcc = 'phs000001.v3.p1')
}
}
