% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_public.R
\docType{methods}
\name{prjConfig}
\alias{prjConfig}
\alias{prjConfig,Commons,character-method}
\title{Create user project directory}
\usage{
prjConfig(object, prjDir)

\S4method{prjConfig}{Commons,character}(object, prjDir)
}
\arguments{
\item{object}{Commons class object.}

\item{prjDir}{a character string. The path to designated project directory.}
}
\value{
a logical value. TRUE if the project directory is created and FALSE if not.
}
\description{
The method creates a project directory and saves the path information of the prject directory in a project config file (~/ncbi/dbgapr_conf/prject_config.json). The project directory is where all the data and metadata files are located. Multiple project directories can be created, but only one is defined as the current. An old project directory can be reinstated to become current again.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,prjDir = character}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
prjConfig(c, prjDir = "/home/user/my_project")
}
}
