% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_public.R
\docType{methods}
\name{recordPrjFileInfo}
\alias{recordPrjFileInfo}
\alias{recordPrjFileInfo,Commons-method}
\title{Extract and save phenotype file meta-info}
\usage{
recordPrjFileInfo(object, ...)

\S4method{recordPrjFileInfo}{Commons}(object, ..., phsAcc = "")
}
\arguments{
\item{object}{Commons class object}

\item{...}{There are optional arguments.}

\item{phsAcc}{a character string. The dbGaP study accession.}
}
\value{
(invisible) a data frame. Meta-info of the dataset files under the prject directory.
}
\description{
The method looks for phenotype files in the project directory, extracts, and saves the file attribute information to a file. The phenotype data files should be available under the project directory before calling this function. Checkout \code{\link{searchCopyPhenoFiles}} for more information if not. The resulting project file information is saved in ~/ncbi/dbgap_conf/pht_file_info_indiv.json. When a study accession is provided, it processes only the files of specified study.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: 
}}

\examples{
\dontrun{

c <- Commons()
recordPrjFileInfo(c)
# or 
recordPrjFileInfo(c, phsAcc="phs000001.v3.p1")
}
}
