% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{saveGapPlot}
\alias{saveGapPlot}
\alias{saveGapPlot,Commons-method}
\title{(internal) Save dbGaP data plot to file}
\usage{
saveGapPlot(object, ...)

\S4method{saveGapPlot}{Commons}(object, ..., plotObj, plotType, phvAccNameCombo,
  saveToDir = "", showPlot = T)
}
\arguments{
\item{object}{Commons class object.}

\item{...}{There are optional arguments.}

\item{plotObj}{a ggplot image object.  The resulting image object from a ggplot function all.}

\item{plotType}{a character string. Type of the plot with the possible value of 'boxplot', 'scatterplot', or 'histogram', etc.}

\item{phvAccNameCombo}{a character string. The concatenated string of a variable accession and name. For example, "phv00053747.v2_LNUCSCORE_phv00053764.v2_LPSC". It is used for naming the saved image files.}

\item{saveToDir}{a character string. (optional) The path to the directory where the plot PDF file is saved. If not provided, the file is saved in the 'temp' directory under the user project directory.}

\item{showPlot}{a logical value. (optional) If TRUE (default), shows the created graph; If FALSE, not show the created graph.  The created graph is saved as files regardless.}
}
\value{
a character list. The full paths to the saved plot files
}
\description{
The method takes the ggplot graph object as an input, and save the graph to PDF and PNG files.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: 
}}

\examples{
\dontrun{

c <- Commons()
accNameCombo = 'phv00053747.v2_LNUCSCORE_phv00053764.v2_LPSC'
saveGapPlot(c, plotObj, plotType = 'boxplot', phvAccNameCombo = accNameCombo) 
}
}
\keyword{internal}
