% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commons.R
\docType{methods}
\name{show,Commons-method}
\alias{show,Commons-method}
\title{Display all class Commons methods}
\usage{
\S4method{show}{Commons}(object)
}
\arguments{
\item{object}{Commons class object.}
}
\description{
The method displays all the methods available in the class Commons.
}
\examples{
\dontrun{

c <- Commons()
show(c)
}
}
