% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\docType{methods}
\name{show,Study-method}
\alias{show,Study-method}
\title{Display class methods}
\usage{
\S4method{show}{Study}(object)
}
\arguments{
\item{object}{Study class object.}
}
\description{
The method displays all the available methods in the Study class.
}
\examples{
\dontrun{

s <- Study(phsAcc = 'phs000001.v1.p1')
show(s) 
}
}
