% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\docType{class}
\name{Study-class}
\alias{Study-class}
\alias{Study}
\title{Class Study}
\description{
The class contains the functions that handle the data of a specified study.
}
\section{Slots}{

\describe{
\item{\code{phsAcc}}{a character string. A dbGaP study accession.}
}}

\examples{
\dontrun{

s <- new("Study", phsAcc = 'phs000001.v1.p1')
s <- Study(phsAcc = 'phs000001.v1.p1')
}
}
