% Generated by hand
\docType{methods}
\name{study_internal}
\alias{study_internal}
\title{Method container}
\description{
    A container of the internal methods of class Study. 
}
\keyword{internal}

\section{Methods (internal, class, Study)}{
\itemize{
    \item \code{\link{checkPhvAccList}}
    \item \code{\link{checkObjStudyByAcc}}
}}
