% Generated by hand
\docType{methods}
\name{study_public}
\alias{study_public}
\title{Method container}
\description{
    A container funtion that hosts the methods of class Study.
}
\keyword{internal}

\section{Methods (public, class, Study)}{
\itemize{
    \item \code{\link{getStudyDatasetInfo}}
    \item \code{\link{getStudyVariableInfo}}
    \item \code{\link{getStudyVariableInfoByTerms}}
    \item \code{\link{getStudyVariableData}}
    \item \code{\link{getIdInfo}}
    \item \code{\link{variableSummary}}
    \item \code{\link{variableBoxplot}}
    \item \code{\link{variableScatterplot}}
    \item \code{\link{variableHistogram}}
    \item \code{\link{variablePiechart}}
    \item \code{\link{variableVenndiagram}}
    \item \code{\link{variableCorrelHeatmap}}
}}
