% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_public.R
\docType{methods}
\name{variableBoxplot}
\alias{variableBoxplot}
\alias{variableBoxplot,Study,character-method}
\title{Variable boxplot}
\usage{
variableBoxplot(object, numPhvAcc, ...)

\S4method{variableBoxplot}{Study,character}(object, numPhvAcc, ...,
  catPhvAcc = "", saveToDir = "", showPlot = T)
}
\arguments{
\item{object}{Study class object.}

\item{numPhvAcc}{a character string. The dbGaP accession of a numeric variable.}

\item{...}{There are optional arguments.}

\item{catPhvAcc}{a character string. (optional) The dbGaP accession of a categorical variable.}

\item{saveToDir}{a character string. (optional) The path to the directory where the plot PDF file is saved. If not provided, the file is saved in the 'temp' directory under the user project directory.}

\item{showPlot}{a logical value. (optional) If TRUE (default), shows the created graph; Not show if FALSE. The graph is always saved in PDF and PNG files regardless shown or not.}
}
\value{
a data frame. The data used for plotting.
}
\description{
The method draws the box-and-whisker plot given the accession of either a numeric variable alone or a numeric plus a categorical variable. The boxplot is drawn with the numeric variable as the function of each category of the categorical variable, respectively. When categorical variable accession is not provided, the boxplot is drawn with the numeric variable as the function of the respective subject indices. The resulting graph is saved as PDF and PNG files.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Study,numPhvAcc = character}: of class Study
}}

\examples{
\dontrun{

s <- Study(phsAcc = 'phs000001.v3.p1')
variableBoxplot(s, numPhvAcc = 'phv00000027.v2', catPhvAcc = 'phv00000032.v2')
}
}
