% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_public.R
\docType{methods}
\name{variableCorrelHeatmap}
\alias{variableCorrelHeatmap}
\alias{variableCorrelHeatmap,Study,character-method}
\title{Variable correleation heatmap}
\usage{
variableCorrelHeatmap(object, catPhvAcc, ...)

\S4method{variableCorrelHeatmap}{Study,character}(object, catPhvAcc = "", ...,
  numPhvAccList = vector(), saveToDir = "", showPlot = T)
}
\arguments{
\item{object}{Study class object.}

\item{catPhvAcc}{a character string. The dbGaP accession of a categorical variable.}

\item{...}{There are optional arguments.}

\item{numPhvAccList}{a character vector. A list of the dbGaP accessions of numerical variables.}

\item{saveToDir}{a character string. (optional) The path to the directory where the plot PDF file is saved. If not provided, the file is saved in the 'temp' directory under the user project directory.}

\item{showPlot}{a logical value. (optional) If TRUE (default), shows the created graph; Not show if FALSE.}
}
\value{
a data frame. The data used for plotting.
}
\description{
The method draws the correlation graphs between different categories of a categorical variable. The resulting graph is saved as PDF and PNG files.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Study,catPhvAcc = character}: of class Study
}}

\examples{
\dontrun{

s <- Study(phsAcc = 'phs000001.v3.p1')
catAcc = 'phv00000031.v2'
numAccList = c('phv00000007.v2', 'phv00053794.v2', 'phv00053786.v2', 'phv00000106.v2') 
variableCorrelHeatmap(s, catPhvAcc = catAcc, numPhvAccList = numAccLlist, showPlot=T)
} 
}
