% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_public.R
\docType{methods}
\name{variablePiechart}
\alias{variablePiechart}
\alias{variablePiechart,Study,character-method}
\title{Categorical variable subject count pie-chart}
\usage{
variablePiechart(object, catPhvAcc, ...)

\S4method{variablePiechart}{Study,character}(object, catPhvAcc = "", ...,
  saveToDir = "", showPlot = T)
}
\arguments{
\item{object}{Study class object.}

\item{catPhvAcc}{a character string. The dbGaP accession of a categorical variable.}

\item{...}{There are optional arguments.}

\item{saveToDir}{a character string. (optional) The path to the directory where the plot PDF file is saved. If not provided, the file is saved in the 'temp' directory under the user project directory.}

\item{showPlot}{a logical value. (optional) If TRUE (default), shows the created graph; Not show if FALSE.}
}
\value{
a data frame. The data used for plotting.
}
\description{
The method draws the pie-chart of subject counts among categories of a categorical variable. The resulting graph is saved as PDF and PNG files.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Study,catPhvAcc = character}: of class Study
}}

\examples{
\dontrun{

s <- Study(phsAcc = 'phs000001.v3.p1')
variablePiechart(s, catPhvAcc = 'phv00053764.v2')
}
}
