% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_public.R
\docType{methods}
\name{variableSummary}
\alias{variableSummary}
\alias{variableSummary,Study,character-method}
\title{Variable data summary}
\usage{
variableSummary(object, phvAcc, ...)

\S4method{variableSummary}{Study,character}(object, phvAcc, ...,
  emptyToNa = FALSE)
}
\arguments{
\item{object}{Study class object.}

\item{phvAcc}{a character string. The dbGaP variable accessions}

\item{...}{There are optional arguments.}

\item{emptyToNa}{a logical value. (optional) If TRUE,  converts the empty value to NA. Not convert if FALSE (default).}
}
\value{
a data frame. The statistical summary data.
}
\description{
The method returns the statistical summary of a given variable of the class study. The statistical summary is also displayed along with the variable meta-info.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Study,phvAcc = character}: of class Study
}}

\examples{
\dontrun{

s <- Study(phsAcc = 'phs000001.v3.p1')
variableSummary(s, phvAcc = 'phv00054119.v1.p1.c2')
}
}
