% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{viewAllDatasetInfo}
\alias{viewAllDatasetInfo}
\alias{viewAllDatasetInfo,Commons-method}
\title{(internal) Display meta-info of all datasets}
\usage{
viewAllDatasetInfo(object, ...)

\S4method{viewAllDatasetInfo}{Commons}(object, ..., showAs = "table",
  editor = "")
}
\arguments{
\item{object}{Commons class object.}

\item{...}{There are optional arguments.}

\item{showAs}{a character string. (optional) When the value is 'table', displays the data as a table through a platform specific table viewer; When it is 'json', displays the json text through a plain text editor; When it is 'text', displays in a brief left-justified text format.}

\item{editor}{a character string. (optional) The name of your favorite plain text editor. It should be executable from a command-line prompt of the respective platform. For example, notepad (Windows), vim, emacs (Unix), gedit (Ubuntu), nedit (CentOS), etc.}
}
\description{
The method displays the meta-info of all datasets available in the project directory. The display is through a separate a table viewer or a plain text editor intended for human to browse and search. Before calling this function, the files under the 'combo_dump' sub-directory of the project directory should be made available through the function getAllDatasetInfo.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
viewAllDatasetInfo(c) 
viewAllDatasetInfo(c, showAs = 'json', editor='notepad')
viewAllDatasetInfo(c, showAs = 'text', editor='gedit')
}
}
\keyword{internal}
