% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_internal.R
\docType{methods}
\name{viewStudyVariableInfo}
\alias{viewStudyVariableInfo}
\alias{viewStudyVariableInfo,Study-method}
\title{(internal) View study variable meta-info}
\usage{
viewStudyVariableInfo(object, ...)

\S4method{viewStudyVariableInfo}{Study}(object, ..., phtAcc = "",
  showAs = "table", editor = "")
}
\arguments{
\item{object}{Study class object.}

\item{...}{There are optional arguments.}

\item{phtAcc}{a character string. The dbGaP dataset accession.}

\item{showAs}{a character string. (optional) When the value is 'table', displays the data as a table through a platform specific table viewer; When it is 'json', displays the json text through a plain text editor; When it is 'text', displays in a brief left-justified text format.}

\item{editor}{a character string. (optional) The name of your favorite plain text editor. It should be executable from a command-line prompt of the respective platform. For example, notepad (Windows), vim, emacs (Unix), gedit (Ubuntu), nedit (CentOS), etc.}
}
\description{
The method displays the meta-info of the variables of either a specified dataset or all variables under the study.
}
\section{Methods (by class)}{
\itemize{
\item \code{Study}: A method of class Study
}}

\examples{
\dontrun{

s <- Study(phsAcc = 'phs000001.v3.p1')
viewStudyVariableInfo(s, phtAcc='pht000373.v2', showAs='json')
}
}
\keyword{internal}
