% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{writeFileInfoJson}
\alias{writeFileInfoJson}
\alias{writeFileInfoJson,Commons-method}
\title{(internal) Write file info json}
\usage{
writeFileInfoJson(object, dbgapFile)

\S4method{writeFileInfoJson}{Commons}(object, dbgapFile)
}
\arguments{
\item{object}{Commons class object}

\item{dbgapFile}{a character string. Decrypted dbGaP phenotype file name with path.}
}
\value{
TRUE a logical value. TRUE if json file is written successfully. FALSE if not.
}
\description{
The method takes the attribute set of a phenotype file, converts it to a data frame, and saves it a json file, which is composed through \code{\link{composeFileInfoJson}}. (Developer's notes: This function only appends a new item to the existing json file. It does not deal with replacing existing items. That is because the existing json file (pht_file_info_indiv.json) is removed and recreated every time when it is called by \code{\link{recordPrjFileInfo}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: of class Commons
}}

\examples{
\dontrun{

c <- Commons()
file = '/home/foo/data/phs000429.v1.pht002481.v1.p1.c1.areds_data_final_11.EDO.txt.gz' 
writeFileInfoJson(c, dbgapFile = file) 
}
}
\keyword{internal}
