% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{writeLog}
\alias{writeLog}
\alias{writeLog,Commons,character-method}
\title{(internal) Error Logging}
\usage{
writeLog(object, type, level, message, show, ...)

\S4method{writeLog}{Commons,character}(object, type, level = "info", message,
  show = FALSE, ..., maxSizeInMb = 1, keptDay = 90)
}
\arguments{
\item{object}{Commons class object.}

\item{type}{a character string. The logging type of either 'setup' for project_setup.log or 'process' for data_process.log.}

\item{level}{a character string. The logging level of either 'error', 'warn', 'info', or 'debug'. The default value is 'info'.}

\item{message}{a character string. The logging message.}

\item{show}{a logical value. If TRUE, console display the log messaage; If FALSE (default), not display.}

\item{...}{There are optional arguments.}

\item{maxSizeInMb}{a integer. (optional) The maximum size (MB) of a log file. The log file is moved to archieved_log directory after reaching the maximum size (default 1MB).}

\item{keptDay}{an integer. (optional) The maximum days of an archieved log file is kept. The archived log file is deleted after the duration Day is past (default 90 days).}
}
\description{
The method writes and displaye logging message to the log files.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,type = character}: of classs Commons
}}

\keyword{internal}
