% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_internal.R
\docType{methods}
\name{checkObjStudyByAcc}
\alias{checkObjStudyByAcc}
\alias{checkObjStudyByAcc,Study,character-method}
\title{(internal) Get study of a dataset or variable}
\usage{
checkObjStudyByAcc(object, acc)

\S4method{checkObjStudyByAcc}{Study,character}(object, acc)
}
\arguments{
\item{object}{Study class object.}

\item{acc}{a character string. A dataset or variable accession.}
}
\value{
a character string. (invisible) The study accession to which the input dataset or variable belongs.
}
\description{
The method checks a given dataset or variable accession to see if it belongs to the class study or not.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Study,acc = character}: A method of class Study
}}

\examples{
\dontrun{ 

s <- Study(phsAcc = 'phs000001.v3.p1')
phsAcc <- checkObjStudyByAcc(s, acc = 'pht000370.v2.p1')
phsAcc <-checkObjStudyByAcc(s, acc = 'phv00054119.v1.p1')
}
}
\keyword{internal}
