% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commons.R
\docType{class}
\name{Commons-class}
\alias{Commons-class}
\alias{Commons}
\title{Base class of dbgapr package}
\description{
The class contains the methods commonly used by all classes and methods in the package. The related functions include setting up the user project directory, copying the decrypted dbGaP phenotype data (datasets) files to the project directory, ftp download of the meta-data (data dictionary) files, merging the datasets split into different consents groups, reviewing and retrieving the data and meta-data by accessions, etc.
}
\examples{
\dontrun{

c <- Commons()
}
}
