% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_internal.R
\docType{methods}
\name{filterBySubjIds}
\alias{filterBySubjIds}
\alias{filterBySubjIds,Commons,data.frame-method}
\title{(internal) Filter dataset data by subject ids}
\usage{
filterBySubjIds(object, varDF, ...)

\S4method{filterBySubjIds}{Commons,data.frame}(object, varDF, ...,
  dbgapIdsOrFile = NULL)
}
\arguments{
\item{object}{Commons class object.}

\item{varDF}{a data frame. The dataset data.}

\item{...}{There are optional arguments.}

\item{dbgapIdsOrFile}{a character vector or a character string. (optional) This argument can be either a vector of ID list or a path to a file that contains a list of IDs. The IDs can be dbGaP_Subject_ID or dbGaP_Sample_ID denpending on type of the data. When the list of IDs is provided by a file, it should be  a plain text file with one ID per line.}
}
\value{
a data frame. Filtered dataset data.
}
\description{
The method filters out the rows of a dataset by a given list of subject ids. The subject ids can be provided as a list or as a text file.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Commons,varDF = data.frame}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
filterBySubjIds(c, varDF = varDF, dbgapIdsOrFile = c("219", "220", "223"))
}
}
\keyword{internal}
