% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_public.R
\docType{methods}
\name{ftpDownload}
\alias{ftpDownload}
\alias{ftpDownload,Commons-method}
\title{Download supplemental data files from dbGaP ftp}
\usage{
ftpDownload(object, ...)

\S4method{ftpDownload}{Commons}(object, ..., phsAcc = "",
  dataStudyOnly = TRUE, overwrite = FALSE)
}
\arguments{
\item{object}{Commons class object}

\item{...}{There are optional arguments.}

\item{phsAcc}{a character string. The dbGaP study accession. Only the files of the given study are downloaded when a study accession is provided.}

\item{dataStudyOnly}{a logical value. When TRUE (default), only downloads the dataset and variable metadata of the stdudies that have data files in the project directory.  When FALSE, downloads the dataset and variable metadata of all dbGaP released studies, regardless the actual phenotype data files of the studies are downloaded or not.}

\item{overwrite}{a logical value. When TRUE, downloads the supplemental metadata files if they aren't already downloaded. When FALSE (default, downloads the metadata files even the they already exist.}
}
\value{
a logical value. If TRUE, all files are downloaded OK. If FALSE, there is at least one failed download.
}
\description{
The method downloads the supplemental metadata files from the dbGaP ftp site. It is called by the \code{\link{prepareData}} function.
}
\section{Methods (by class)}{
\itemize{
\item \code{Commons}: A method of class Commons
}}

\examples{
\dontrun{

c <- Commons()
ftpDownload(c)
# or
ftpDownload(c, phsAcc = 'phs000001.v3.p1')
ftpDownload(c, dataStudyOnly=FALSE)
# or to include the data meta-info of all the studies released in dbGaP.
# Note: This step may take serveral hours.
ftpDownload(c, dataStudyOnly = FALSE)
}
}
