% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_public.R
\docType{methods}
\name{getStudyVariableData}
\alias{getStudyVariableData}
\alias{getStudyVariableData,Study-method}
\title{Get study variable data by accession}
\usage{
getStudyVariableData(object, ...)

\S4method{getStudyVariableData}{Study}(object, ..., phtAcc = "",
  phvAccList = vector(), subjIdsOrFile = NA, emptyToNa = FALSE,
  colNameWithAcc = FALSE)
}
\arguments{
\item{object}{Study class object.}

\item{...}{There are optional arguments.}

\item{phtAcc}{a character string. (optional) The dbGaP phenotype dataset accession.}

\item{phvAccList}{a character vector. (optional) The dbGaP variable accessions.}

\item{subjIdsOrFile}{a character vector or a character string. (optional) This argument is either list of dbGaP subject ids (dbGaP_Subject_ID) of the variables or the path to a file that contains a list of the dbGaP subject ids of the variables. The file is a plain text file with one dbGaP subject id per line.}

\item{emptyToNa}{a logical value. (optional). If TRUE, converts the empty values to NA; If FALSE (default), not convert.}

\item{colNameWithAcc}{a logical value. (optional) If TRUE, the variable column name is concatenated with the respective variable accession (e.g. AGEPHOT_phv00000027.v2); If FALSE (default), keep the original column name unchanged (e.g. AGEPHOT).}
}
\value{
a data frame. Merged data of the input variables.
}
\description{
The method retrieves the variable data given a list of dbGaP variable accessions or a dataset accession.
}
\section{Methods (by class)}{
\itemize{
\item \code{Study}: of class Study
}}

\examples{
\dontrun{

s <- Study(phsAcc = 'phs000001.v3.p1')
var_list <- c('phv00054119.v1.p1.c2', 'phv00054118.v1.p1', 'phv00053733.v2')
getStudyVariableData(s, phvAccList = var_list)
# or
getStudyVariableData(s, phtAcc = 'pht000370.v2.p1', colNameWithAcc = TRUE)
}
}
