% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_internal.R
\docType{methods}
\name{getStudyVariableInfoByDataType}
\alias{getStudyVariableInfoByDataType}
\alias{getStudyVariableInfoByDataType,Study,character-method}
\title{(internal) Get meta-info of study numerical or categorical variables}
\usage{
getStudyVariableInfoByDataType(object, dataType, ...)

\S4method{getStudyVariableInfoByDataType}{Study,character}(object, dataType,
  ..., dataDicDF = data.frame())
}
\arguments{
\item{object}{Study class object.}

\item{dataType}{a character string. The variable datatype. The possible value is either 'num' (for numerical variable) or 'cat' (for categorical variable).}

\item{...}{There are optional arguments.}

\item{dataDicDF}{a data frame. (optional) Dataset data dictionary data.}
}
\value{
a data frame. The variable meta-info.
}
\description{
The method returns the variable meta-info of either numerical or categorical variables of the study.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Study,dataType = character}: A method of class Study
}}

\examples{
\dontrun{

getStudyVariableInfoByDataType(s, dataType = 'num', dataDicDF = dataDicDF)
getStudyVariableInfoByDataType(s, dataType = 'cat', dataDicDF = dataDicDF)
}
}
\keyword{internal}
