% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_internal.R
\docType{methods}
\name{getVariableDataByPhvAccAndSubjId}
\alias{getVariableDataByPhvAccAndSubjId}
\alias{getVariableDataByPhvAccAndSubjId,Study,character-method}
\title{(internal) Get variable data subset by subject ids}
\usage{
getVariableDataByPhvAccAndSubjId(object, phvAccList, ...)

\S4method{getVariableDataByPhvAccAndSubjId}{Study,character}(object, phvAccList,
  ..., dbgapIdsOrFile = NULL)
}
\arguments{
\item{object}{Study class object}

\item{phvAccList}{a character vector. A list of dbGaP variable accessions.}

\item{...}{There are optional arguments.}

\item{dbgapIdsOrFile}{a character vector or a character string. (optional) This argument can be either a vector of ID list or a path to a file that contains a list of IDs. The IDs can be dbGaP_Subject_ID or dbGaP_Sample_ID denpending on type of the data. When the list of IDs is provided by a file, it should be  a plain text file with one ID per line.}
}
\value{
a data frame. The variable data of given variables subset by the given subjects.
}
\description{
The method returns the data of a given list of variables subset by a list of subject ids.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Study,phvAccList = character}: returns a dataframe of the variable data
}}

\examples{
\dontrun{

s <- Study(phsAcc = 'phs000001.v3.p1')
ids <- c("219", "220", "221") 
accList <- c('phv00054119.v1.p1.c2', 'phv00053735.v2')
getVariableDataByPhvAccAndSubjId(s, phvAccList = accList, dbgapIdsOrFile = ids)
idFile = '/home/user/temp/subj_ids.txt'
getVariableDataByPhvAccAndSubjId(s, phvAccList = accList, dbgapIdsOrFile = idFile)
}
}
\keyword{internal}
