% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\docType{class}
\name{Study-class}
\alias{Study-class}
\alias{Study}
\title{Class Study}
\description{
The class contains the methods that handle the data of a specified study. The functions of this class include viewing and retrieving the data and meta-data of the specified study by accessions, visulizing variable data through a variety of plotting methods, etc. The data (datasets, variables) and meta-data (data ditionary) files of the study need to be available under the user project directory before calling the functions under this class. A dbGaP study accession is rerquired for initializing the class.
}
\section{Slots}{

\describe{
\item{\code{phsAcc}}{a character string. A dbGaP study accession.}
}}

\examples{
\dontrun{

s <- new("Study", phsAcc = 'phs000001.v1.p1')
s <- Study(phsAcc = 'phs000001.v1.p1')
}
}
